refactor: Update LibraryScreen font and add settings button
This commit updates the font used for the "Library" title in `LibraryScreen.kt` from `ExpTitleTypography.titleMedium` (which uses Montserrat) to `GoogleSansRounded`. The font size has also been increased to `40.sp`.

Additionally, a `FilledIconButton` for accessing settings has been added to the `actions` section of the `TopAppBar`. This button navigates to the `Screen.Settings.route` when clicked.