refactor: Clean up NavBarStyle imports
This commit removes unused imports of `NavBarStyle` from several files and organizes existing ones.

It also introduces `NavBarContentHeightFullWidth` in `PlayerInternalNavigationBar.kt` to define a specific height for the navigation bar when `navBarStyle` is `FULL_WIDTH`. This value is then used to set the height of the `Box` composable within `PlayerInternalNavigationBar`.

In `SettingsScreen.kt`, the `Spacer` preceding the "NavBar Corner Radius" setting is now conditionally rendered, only appearing if `navBarStyle` is `DEFAULT`.